requirements.txt
pymupdf==1.22.5
openai==0.27.8
tiktoken==0.4.0
qdrant_client==1.3.2
gradio==3.39.0
langchain==0.0.249
pydantic==1.10.12
python-dotenv==1.0.0
tqdm==4.65.0